move global initialization out of core.js
