+++ src/env.js
@@ -22,15 +22,6 @@ Object.assign(esmsInitOptions, self.esmsInitOptions || {});
 
 export const version = self.VERSION;
 
-const r = esmsInitOptions.version;
-if (self.importShim || (r && r !== version)) {
-  if (self.ESMS_DEBUG)
-    console.info(
-      `es-module-shims: skipping initialization as ${r ? `configured for ${r}` : 'another instance has already registered'}`
-    );
-  $ret();
-}
-
 // shim mode is determined on initialization, no late shim mode
 export const shimMode =
   esmsInitOptions.shimMode ||
